jobcraft/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
python-multipart==0.0.19
google-generativeai>=0.8.0
anthropic==0.42.0
openai>=1.55.0
PyPDF2==3.0.1
python-docx==1.1.2
reportlab==4.2.5
playwright==1.49.1
pydantic==2.10.3
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
